Update to Snort 2.9.4.

+++ src/detection-plugins/sp_flowbits.h
@@ -16,7 +16,7 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  *
  ****************************************************************************/
  
@@ -54,7 +54,7 @@ void FlowbitResetCounts(void);
 */
 typedef struct _FLOWBITS_OBJECT
 {
-    uint16_t id;
+    uint32_t id;
     uint8_t  types;
     int toggle;
     int set;
